refactor!: revert 7076fc2ae987948e2c52b696666842ddb67af5d7
hoist ObservableWallet to cip30 WalletApi mapping back to 'wallet' package
