+++ packages/cip30/src/WalletApi/index.ts
@@ -1,3 +1,2 @@
 export * from './Cip30Wallet';
 export * from './types';
-export * from './observableWalletApi';
