+++ packages/cip30/src/tsconfig.json
@@ -7,9 +7,6 @@
   "references": [
     {
       "path": "../../core/src"
-    },
-    {
-      "path": "../../wallet/src"
     }
   ]
 }
