+++ packages/wallet/package.json
@@ -30,7 +30,6 @@
     "@cardano-ogmios/client": "^5.2.0",
     "@cardano-sdk/cardano-graphql": " 0.2.0",
     "@cardano-sdk/ogmios": " 0.2.0",
-    "@cardano-sdk/cip30": " 0.2.0",
     "@cardano-sdk/util-dev": " 0.2.0",
     "@types/node-hid": "^1.3.1",
     "@types/pbkdf2": "^3.1.0",
@@ -47,6 +46,7 @@
   "dependencies": {
     "@cardano-foundation/ledgerjs-hw-app-cardano": "5.0.0",
     "@cardano-sdk/cip2": " 0.2.0",
+    "@cardano-sdk/cip30": " 0.2.0",
     "@cardano-sdk/core": " 0.2.0",
     "@emurgo/cardano-message-signing-nodejs": "^1.0.1",
     "@ledgerhq/hw-transport-node-hid-noevents": "6.27.1",
