+++ packages/wallet/src/SingleAddressWallet.ts
@@ -51,13 +51,14 @@ import {
   distinctBlock,
   distinctTimeSettings
 } from './services';
-import { Cip30DataSignature, cip30signData } from './KeyManagement/cip8';
+import { Cip30DataSignature } from '@cardano-sdk/cip30';
 import { InputSelector, defaultSelectionConstraints, roundRobinRandomImprove } from '@cardano-sdk/cip2';
 import { Logger, dummyLogger } from 'ts-log';
 import { Observable, Subject, combineLatest, filter, lastValueFrom, map, take } from 'rxjs';
 import { RetryBackoffConfig } from 'backoff-rxjs';
 import { TxInternals, createTransactionInternals, ensureValidityInterval } from './Transaction';
 import { WalletStores, createInMemoryWalletStores } from './persistence';
+import { cip30signData } from './KeyManagement/cip8';
 import { isEqual } from 'lodash-es';
 
 export interface SingleAddressWalletProps {
