+++ packages/wallet/src/index.ts
@@ -4,3 +4,4 @@ export * from './SingleAddressWallet';
 export * from './types';
 export * from './services';
 export * as storage from './persistence';
+export * as cip30 from './cip30';
